+++ app/reactstrap.d.ts
@@ -1,6 +1,7 @@
 /* tslint:disable */
 import * as React from 'react';
 import { UncontrolledDropdownProps } from 'reactstrap/lib/Dropdown';
+import { DropdownToggleProps } from 'reactstrap/lib/DropdownToggle';
 import { FadeProps } from 'reactstrap/lib/Fade';
 
 declare module 'reactstrap/lib/Dropdown' {
@@ -8,6 +9,12 @@ declare module 'reactstrap/lib/Dropdown' {
     dropup?: boolean;
     size?: string;
     tag?: React.ReactType;
+  }
+}
+
+declare module 'reactstrap/lib/DropdownToggle' {
+  interface DropdownToggleProps {
+    href?: string;
   }
 }